imported xdvik-22.84.15
git-svn-id: svn://tug.org/texlive/trunk/Build/source@12586 c570f23f-e606-0410-a88d-b1316a301751

+++ texk/xdvik/gui/Tip.c
@@ -131,10 +131,10 @@ static XtResource resources[] = {
      XtRString, "fixed"},
     {XmNwaitPeriod, XmCWaitPeriod,
      XtRInt, sizeof(int), offset(tip.waitPeriod),
-     XtRImmediate, (XtPointer) 800},
+     XtRImmediate, (XtPointer)(ptrdiff_t)800},
     {XmNcancelWaitPeriod, XmCCancelWaitPeriod,
      XtRInt, sizeof(int), offset(tip.cwp),
-     XtRImmediate, (XtPointer) 250},
+     XtRImmediate, (XtPointer)(ptrdiff_t)250},
 };
 #undef offset
